requirements.txt
python-telegram-bot==20.7
uvicorn==0.30.6
starlette==0.37.2